refactor: share-option-items.ts 폴더 이동, kakaoshare 문구 변경
